refactor(server): extract backup + notification route groups (Refs #3502)
Phase 2 Batch 2 of the server.ts route extraction (#3502). Moves four
cohesive inline route groups out of the server.ts composition root into
dedicated routes/ modules, behaviour-preserving (handler bodies moved
verbatim, auth/validation middleware unchanged).

Extracted:
- /backup/*        -> backupRoutes.ts (backupRouter)
- /system/backup/* -> backupRoutes.ts (systemBackupRouter)
- /push/*          -> notificationRoutes.ts (pushRouter)
- /apprise/*       -> notificationRoutes.ts (appriseRouter)

Notification helpers (getUserNotificationPreferencesAsync,
saveUserNotificationPreferencesAsync, applyNodeNamePrefixAsync) were
already importable from utils/notificationFiltering.js, so no helper had
to move; the test/Apprise handlers keep using the meshtasticManager
default-export singleton exactly as before. Orphaned server.ts imports
(systemBackupService + the three notification helpers) removed.

server.ts: 154 -> 130 inline apiRouter handlers (-24); ~10,476 -> ~9,449 LOC.

Adds backupRoutes.test.ts and notificationRoutes.test.ts (39 new tests).
Full Vitest suite green (7261 tests, 0 failures); tsc clean.
