+++ src/server/server.ts
@@ -37,7 +37,6 @@ import { backupSchedulerService } from './services/backupSchedulerService.js';
 import { databaseMaintenanceService } from './services/databaseMaintenanceService.js';
 import { positionEstimationScheduler } from './services/positionEstimationScheduler.js';
 import { autoFavoriteManagementScheduler } from './services/autoFavoriteManagementService.js';
-import { systemBackupService } from './services/systemBackupService.js';
 import { systemRestoreService } from './services/systemRestoreService.js';
 import { duplicateKeySchedulerService } from './services/duplicateKeySchedulerService.js';
 import { waypointRebroadcastSchedulerService } from './services/waypointRebroadcastSchedulerService.js';
@@ -48,7 +47,6 @@ import { inactiveNodeNotificationService } from './services/inactiveNodeNotifica
 import { lowBatteryNotificationService } from './services/lowBatteryNotificationService.js';
 import { serverEventNotificationService } from './services/serverEventNotificationService.js';
 import { autoDeleteByDistanceService } from './services/autoDeleteByDistanceService.js';
-import { getUserNotificationPreferencesAsync, saveUserNotificationPreferencesAsync, applyNodeNamePrefixAsync } from './utils/notificationFiltering.js';
 import { upgradeService } from './services/upgradeService.js';
 import { enhanceNodeForClient, filterNodesByChannelPermission, checkNodeChannelAccess, getEffectiveDbNodePosition } from './utils/nodeEnhancer.js';
 import { dynamicCspMiddleware, refreshTileHostnameCache } from './middleware/dynamicCsp.js';
@@ -983,6 +981,8 @@ import routeSegmentRoutes from './routes/routeSegmentRoutes.js';
 import neighborInfoRoutes from './routes/neighborInfoRoutes.js';
 import ignoredNodeRoutes from './routes/ignoredNodeRoutes.js';
 import announceRoutes from './routes/announceRoutes.js';
+import { backupRouter, systemBackupRouter } from './routes/backupRoutes.js';
+import { pushRouter, appriseRouter } from './routes/notificationRoutes.js';
 
 // CSRF token endpoint (must be before CSRF protection middleware)
 apiRouter.get('/csrf-token', csrfTokenEndpoint);
@@ -1107,6 +1107,10 @@ apiRouter.use('/route-segments', routeSegmentRoutes);
 apiRouter.use('/neighbor-info', neighborInfoRoutes);
 apiRouter.use('/ignored-nodes', ignoredNodeRoutes);
 apiRouter.use('/announce', announceRoutes);
+apiRouter.use('/backup', backupRouter);
+apiRouter.use('/system/backup', systemBackupRouter);
+apiRouter.use('/push', pushRouter);
+apiRouter.use('/apprise', appriseRouter);
 
 // Wire up side-effect callbacks for settingsRoutes
 setSettingsCallbacks({
@@ -5297,322 +5301,6 @@ apiRouter.get('/device/backup', requirePermission('configuration', 'read'), asyn
   }
 });
 
-// Get backup settings
-apiRouter.get('/backup/settings', requirePermission('configuration', 'read'), async (_req, res) => {
-  try {
-    const enabled = await databaseService.settings.getSetting('backup_enabled') === 'true';
-    const maxBackups = parseInt(await databaseService.settings.getSetting('backup_maxBackups') || '7', 10);
-    const backupTime = await databaseService.settings.getSetting('backup_time') || '02:00';
-
-    res.json({
-      enabled,
-      maxBackups,
-      backupTime,
-    });
-  } catch (error) {
-    logger.error('❌ Error getting backup settings:', error);
-    res.status(500).json({
-      error: 'Failed to get backup settings',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Save backup settings
-apiRouter.post('/backup/settings', requirePermission('configuration', 'write'), async (req, res) => {
-  try {
-    const { enabled, maxBackups, backupTime } = req.body;
-
-    // Validate inputs
-    if (typeof enabled !== 'boolean') {
-      return res.status(400).json({ error: 'Invalid enabled value' });
-    }
-
-    if (typeof maxBackups !== 'number' || maxBackups < 1 || maxBackups > 365) {
-      return res.status(400).json({ error: 'Invalid maxBackups value (must be 1-365)' });
-    }
-
-    if (!backupTime || !/^\d{2}:\d{2}$/.test(backupTime)) {
-      return res.status(400).json({ error: 'Invalid backupTime format (must be HH:MM)' });
-    }
-
-    // Save settings
-    await databaseService.settings.setSetting('backup_enabled', enabled.toString());
-    await databaseService.settings.setSetting('backup_maxBackups', maxBackups.toString());
-    await databaseService.settings.setSetting('backup_time', backupTime);
-
-    logger.info(`⚙️  Backup settings updated: enabled=${enabled}, maxBackups=${maxBackups}, time=${backupTime}`);
-
-    res.json({ success: true });
-  } catch (error) {
-    logger.error('❌ Error saving backup settings:', error);
-    res.status(500).json({
-      error: 'Failed to save backup settings',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// List all backups
-apiRouter.get('/backup/list', requirePermission('configuration', 'read'), async (_req, res) => {
-  try {
-    const backups = await backupFileService.listBackups();
-    res.json(backups);
-  } catch (error) {
-    logger.error('❌ Error listing backups:', error);
-    res.status(500).json({
-      error: 'Failed to list backups',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Download a specific backup
-apiRouter.get('/backup/download/:filename', requirePermission('configuration', 'read'), async (req, res) => {
-  try {
-    const { filename } = req.params;
-
-    // Validate filename to prevent directory traversal - only allow alphanumeric, hyphens, underscores, and .yaml extension
-    if (!/^[a-zA-Z0-9\-_]+\.yaml$/.test(filename)) {
-      return res.status(400).json({ error: 'Invalid filename format' });
-    }
-
-    const content = await backupFileService.getBackup(filename);
-
-    res.setHeader('Content-Type', 'application/x-yaml');
-    res.setHeader('Content-Disposition', `attachment; filename="${filename}"`);
-    res.send(content);
-
-    logger.info(`📥 Backup downloaded: ${filename}`);
-  } catch (error) {
-    logger.error('❌ Error downloading backup:', error);
-    res.status(500).json({
-      error: 'Failed to download backup',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Delete a specific backup
-apiRouter.delete('/backup/delete/:filename', requirePermission('configuration', 'write'), async (req, res) => {
-  try {
-    const { filename } = req.params;
-
-    // Validate filename to prevent directory traversal - only allow alphanumeric, hyphens, underscores, and .yaml extension
-    if (!/^[a-zA-Z0-9\-_]+\.yaml$/.test(filename)) {
-      return res.status(400).json({ error: 'Invalid filename format' });
-    }
-
-    await backupFileService.deleteBackup(filename);
-
-    logger.info(`🗑️  Backup deleted: ${filename}`);
-
-    res.json({ success: true });
-  } catch (error) {
-    logger.error('❌ Error deleting backup:', error);
-    res.status(500).json({
-      error: 'Failed to delete backup',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// ========== System Backup Endpoints ==========
-
-// Create a system backup (exports all database tables to JSON)
-apiRouter.post('/system/backup', requirePermission('configuration', 'write'), async (req, res) => {
-  try {
-    logger.info('📦 System backup requested...');
-
-    const dirname = await systemBackupService.createBackup('manual');
-
-    // Audit log
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'system_backup_created',
-      'system_backup',
-      JSON.stringify({ dirname, type: 'manual' }),
-      req.ip || null
-    );
-
-    logger.info(`✅ System backup created: ${dirname}`);
-
-    res.json({
-      success: true,
-      dirname,
-      message: 'System backup created successfully',
-    });
-  } catch (error) {
-    logger.error('❌ Error creating system backup:', error);
-    res.status(500).json({
-      error: 'Failed to create system backup',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// List all system backups
-apiRouter.get('/system/backup/list', requirePermission('configuration', 'read'), async (_req, res) => {
-  try {
-    const backups = await systemBackupService.listBackups();
-    res.json(backups);
-  } catch (error) {
-    logger.error('❌ Error listing system backups:', error);
-    res.status(500).json({
-      error: 'Failed to list system backups',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Download a system backup as tar.gz
-apiRouter.get('/system/backup/download/:dirname', requirePermission('configuration', 'read'), async (req, res) => {
-  try {
-    const { dirname } = req.params;
-
-    // Validate dirname to prevent directory traversal - only allow date format YYYY-MM-DD_HHMMSS
-    if (!/^\d{4}-\d{2}-\d{2}_\d{6}$/.test(dirname)) {
-      return res.status(400).json({ error: 'Invalid backup directory name format' });
-    }
-
-    const backupPath = systemBackupService.getBackupPath(dirname);
-    // archiver v8 exposes only named class exports; @types/archiver still ships v7 types.
-    const { TarArchive } = (await import('archiver')) as unknown as {
-      TarArchive: new (opts: import('archiver').ArchiverOptions) => import('archiver').Archiver;
-    };
-    const fs = await import('fs');
-
-    // Check if backup exists
-    if (!fs.existsSync(backupPath)) {
-      return res.status(404).json({ error: 'Backup not found' });
-    }
-
-    // Create tar.gz archive on-the-fly
-    res.setHeader('Content-Type', 'application/gzip');
-    res.setHeader('Content-Disposition', `attachment; filename="${dirname}.tar.gz"`);
-
-    const archive = new TarArchive({
-      gzip: true,
-      gzipOptions: { level: 9 },
-    });
-
-    archive.on('error', err => {
-      logger.error('❌ Error creating archive:', err);
-      res.status(500).json({ error: 'Failed to create archive' });
-    });
-
-    // Audit log before streaming
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'system_backup_downloaded',
-      'system_backup',
-      JSON.stringify({ dirname }),
-      req.ip || null
-    );
-
-    archive.pipe(res);
-    archive.directory(backupPath, dirname);
-    await archive.finalize();
-
-    logger.info(`📥 System backup downloaded: ${dirname}`);
-  } catch (error) {
-    logger.error('❌ Error downloading system backup:', error);
-    res.status(500).json({
-      error: 'Failed to download system backup',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Delete a system backup
-apiRouter.delete('/system/backup/delete/:dirname', requirePermission('configuration', 'write'), async (req, res) => {
-  try {
-    const { dirname } = req.params;
-
-    // Validate dirname to prevent directory traversal
-    if (!/^\d{4}-\d{2}-\d{2}_\d{6}$/.test(dirname)) {
-      return res.status(400).json({ error: 'Invalid backup directory name format' });
-    }
-
-    await systemBackupService.deleteBackup(dirname);
-
-    // Audit log
-    databaseService.auditLogAsync(
-      req.user!.id,
-      'system_backup_deleted',
-      'system_backup',
-      JSON.stringify({ dirname }),
-      req.ip || null
-    );
-
-    logger.info(`🗑️  System backup deleted: ${dirname}`);
-
-    res.json({ success: true });
-  } catch (error) {
-    logger.error('❌ Error deleting system backup:', error);
-    res.status(500).json({
-      error: 'Failed to delete system backup',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Get system backup settings
-apiRouter.get('/system/backup/settings', requirePermission('configuration', 'read'), async (_req, res) => {
-  try {
-    const enabled = await databaseService.settings.getSetting('system_backup_enabled') === 'true';
-    const maxBackups = parseInt(await databaseService.settings.getSetting('system_backup_maxBackups') || '7', 10);
-    const backupTime = await databaseService.settings.getSetting('system_backup_time') || '03:00';
-
-    res.json({
-      enabled,
-      maxBackups,
-      backupTime,
-    });
-  } catch (error) {
-    logger.error('❌ Error getting system backup settings:', error);
-    res.status(500).json({
-      error: 'Failed to get system backup settings',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
-// Save system backup settings
-apiRouter.post('/system/backup/settings', requirePermission('configuration', 'write'), async (req, res) => {
-  try {
-    const { enabled, maxBackups, backupTime } = req.body;
-
-    // Validate inputs
-    if (typeof enabled !== 'boolean') {
-      return res.status(400).json({ error: 'Invalid enabled value' });
-    }
-
-    if (typeof maxBackups !== 'number' || maxBackups < 1 || maxBackups > 365) {
-      return res.status(400).json({ error: 'Invalid maxBackups value (must be 1-365)' });
-    }
-
-    if (!backupTime || !/^\d{2}:\d{2}$/.test(backupTime)) {
-      return res.status(400).json({ error: 'Invalid backupTime format (must be HH:MM)' });
-    }
-
-    // Save settings
-    await databaseService.settings.setSetting('system_backup_enabled', enabled.toString());
-    await databaseService.settings.setSetting('system_backup_maxBackups', maxBackups.toString());
-    await databaseService.settings.setSetting('system_backup_time', backupTime);
-
-    logger.info(`⚙️  System backup settings updated: enabled=${enabled}, maxBackups=${maxBackups}, time=${backupTime}`);
-
-    res.json({ success: true });
-  } catch (error) {
-    logger.error('❌ Error saving system backup settings:', error);
-    res.status(500).json({
-      error: 'Failed to save system backup settings',
-      details: error instanceof Error ? error.message : 'Unknown error',
-    });
-  }
-});
-
 // ==========================================
 // Refresh nodes from device endpoint
 apiRouter.post('/nodes/refresh', requirePermission('nodes', 'write'), async (req, res) => {
@@ -8529,721 +8217,6 @@ apiRouter.post('/system/restart', requirePermission('settings', 'write'), (_req,
   }
 });
 
-// ==========================================
-// Push Notification Endpoints
-// ==========================================
-
-// Get VAPID public key and configuration status
-apiRouter.get('/push/vapid-key', optionalAuth(), async (_req, res) => {
-  const publicKey = await pushNotificationService.getPublicKeyAsync();
-  const status = await pushNotificationService.getVapidStatusAsync();
-
-  res.json({
-    publicKey,
-    status,
-  });
-});
-
-// Get push notification status
-apiRouter.get('/push/status', optionalAuth(), async (_req, res) => {
-  const status = await pushNotificationService.getVapidStatusAsync();
-  res.json(status);
-});
-
-// Update VAPID subject (admin only)
-apiRouter.put('/push/vapid-subject', requireAdmin(), async (req, res) => {
-  try {
-    const { subject } = req.body;
-
-    if (!subject || typeof subject !== 'string') {
-      return res.status(400).json({ error: 'Subject is required and must be a string' });
-    }
-
-    await pushNotificationService.updateVapidSubject(subject);
-    res.json({ success: true, subject });
-  } catch (error: any) {
-    logger.error('Error updating VAPID subject:', error);
-    res.status(400).json({ error: error.message || 'Failed to update VAPID subject' });
-  }
-});
-
-// Subscribe to push notifications
-apiRouter.post(
-  '/push/subscribe',
-  optionalAuth(),
-  requirePermission('messages', 'read', { sourceIdFrom: 'body' }),
-  async (req, res) => {
-    try {
-      const { subscription, sourceId } = req.body;
-
-      if (!subscription || !subscription.endpoint || !subscription.keys) {
-        return res.status(400).json({ error: 'Invalid subscription data' });
-      }
-      if (!sourceId || typeof sourceId !== 'string') {
-        return res.status(400).json({ error: 'sourceId is required' });
-      }
-
-      // Validate source exists
-      const source = await databaseService.sources.getSource(sourceId);
-      if (!source) {
-        return res.status(400).json({ error: `Unknown sourceId: ${sourceId}` });
-      }
-
-      const userId = req.session?.userId;
-      const userAgent = req.headers['user-agent'];
-
-      await pushNotificationService.saveSubscription(userId, subscription, userAgent, sourceId);
-
-      res.json({ success: true });
-    } catch (error: any) {
-      logger.error('Error saving push subscription:', error);
-      res.status(500).json({ error: error.message || 'Failed to save subscription' });
-    }
-  }
-);
-
-// Unsubscribe from push notifications
-apiRouter.post(
-  '/push/unsubscribe',
-  optionalAuth(),
-  requirePermission('messages', 'read', { sourceIdFrom: 'body' }),
-  async (req, res) => {
-    try {
-      const { endpoint, sourceId } = req.body;
-
-      if (!endpoint) {
-        return res.status(400).json({ error: 'Endpoint is required' });
-      }
-      if (!sourceId || typeof sourceId !== 'string') {
-        return res.status(400).json({ error: 'sourceId is required' });
-      }
-
-      await pushNotificationService.removeSubscription(endpoint);
-
-      res.json({ success: true });
-    } catch (error: any) {
-      logger.error('Error removing push subscription:', error);
-      res.status(500).json({ error: error.message || 'Failed to remove subscription' });
-    }
-  }
-);
-
-// Test push notification (admin only)
-apiRouter.post('/push/test', requireAdmin(), async (req, res) => {
-  try {
-    const userId = req.session?.userId;
-
-    // Get local node name for prefix
-    const localNodeInfo = meshtasticManager.getLocalNodeInfo();
-    const localNodeName = localNodeInfo?.longName || null;
-
-    // Apply prefix if user has it enabled
-    const baseBody = 'This is a test push notification from MeshMonitor';
-    const body = await applyNodeNamePrefixAsync(userId, baseBody, localNodeName);
-
-    const result = await pushNotificationService.sendToUser(userId, {
-      title: 'Test Notification',
-      body,
-      icon: '/logo.png',
-      badge: '/logo.png',
-      tag: 'test-notification',
-    });
-
-    res.json({
-      success: true,
-      sent: result.sent,
-      failed: result.failed,
-    });
-  } catch (error: any) {
-    logger.error('Error sending test notification:', error);
-    res.status(500).json({ error: error.message || 'Failed to send test notification' });
-  }
-});
-
-// Get notification preferences (unified for Web Push and Apprise)
-apiRouter.get(
-  '/push/preferences',
-  requireAuth(),
-  requirePermission('messages', 'read', { sourceIdFrom: 'query' }),
-  async (req, res) => {
-  try {
-    const userId = req.session?.userId;
-    if (!userId) {
-      return res.status(401).json({ error: 'Authentication required' });
-    }
-
-    const sourceId = typeof req.query.sourceId === 'string' && req.query.sourceId
-      ? req.query.sourceId
-      : undefined;
-
-    const prefs = await getUserNotificationPreferencesAsync(userId, sourceId);
-
-    if (prefs) {
-      res.json(prefs);
-    } else {
-      // Return defaults
-      res.json({
-        enableWebPush: true,
-        enableApprise: false,
-        enabledChannels: [],
-        enableDirectMessages: true,
-        notifyOnEmoji: true,
-        notifyOnMqtt: true,
-        notifyOnNewNode: true,
-        notifyOnTraceroute: true,
-        notifyOnInactiveNode: false,
-        notifyOnLowBattery: false,
-        lowBatteryThreshold: 20,
-        lowBatteryVoltageThreshold: 3300,
-        notifyOnServerEvents: false,
-        prefixWithNodeName: false,
-        monitoredNodes: [],
-        whitelist: ['Hi', 'Help'],
-        blacklist: ['Test', 'Copy'],
-        appriseUrls: [],
-        mutedChannels: [],
-        mutedDMs: [],
-      });
-    }
-  } catch (error: any) {
-    logger.error('Error loading notification preferences:', error);
-    res.status(500).json({ error: error.message || 'Failed to load preferences' });
-  }
-  }
-);
-
-// Save notification preferences (unified for Web Push and Apprise)
-apiRouter.post(
-  '/push/preferences',
-  requireAuth(),
-  requirePermission('messages', 'read', { sourceIdFrom: 'body' }),
-  async (req, res) => {
-  try {
-    const userId = req.session?.userId;
-    if (!userId) {
-      return res.status(401).json({ error: 'Authentication required' });
-    }
-
-    const sourceId = typeof req.body?.sourceId === 'string' && req.body.sourceId
-      ? req.body.sourceId
-      : undefined;
-
-    const {
-      enableWebPush,
-      enableApprise,
-      enabledChannels,
-      enableDirectMessages,
-      notifyOnEmoji,
-      notifyOnMqtt,
-      notifyOnNewNode,
-      notifyOnTraceroute,
-      notifyOnInactiveNode,
-      notifyOnLowBattery,
-      lowBatteryThreshold,
-      lowBatteryVoltageThreshold,
-      notifyOnServerEvents,
-      prefixWithNodeName,
-      monitoredNodes,
-      whitelist,
-      blacklist,
-      appriseUrls,
-      mutedChannels,
-      mutedDMs,
-    } = req.body;
-
-    // Validate input
-    if (
-      typeof enableWebPush !== 'boolean' ||
-      typeof enableApprise !== 'boolean' ||
-      !Array.isArray(enabledChannels) ||
-      typeof enableDirectMessages !== 'boolean' ||
-      typeof notifyOnEmoji !== 'boolean' ||
-      typeof notifyOnMqtt !== 'boolean' ||
-      typeof notifyOnNewNode !== 'boolean' ||
-      typeof notifyOnTraceroute !== 'boolean' ||
-      typeof notifyOnInactiveNode !== 'boolean' ||
-      typeof notifyOnServerEvents !== 'boolean' ||
-      typeof prefixWithNodeName !== 'boolean' ||
-      !Array.isArray(whitelist) ||
-      !Array.isArray(blacklist)
-    ) {
-      return res.status(400).json({ error: 'Invalid preferences data' });
-    }
-
-    // Validate monitoredNodes is an array of strings
-    if (monitoredNodes !== undefined && !Array.isArray(monitoredNodes)) {
-      return res.status(400).json({ error: 'monitoredNodes must be an array' });
-    }
-
-    // Validate each element is a string
-    if (monitoredNodes && monitoredNodes.some((id: any) => typeof id !== 'string')) {
-      return res.status(400).json({ error: 'monitoredNodes must be an array of strings' });
-    }
-
-    // notifyOnLowBattery / lowBatteryThreshold are optional (older clients omit them)
-    if (notifyOnLowBattery !== undefined && typeof notifyOnLowBattery !== 'boolean') {
-      return res.status(400).json({ error: 'notifyOnLowBattery must be a boolean' });
-    }
-    if (
-      lowBatteryThreshold !== undefined &&
-      (typeof lowBatteryThreshold !== 'number' ||
-        !Number.isFinite(lowBatteryThreshold) ||
-        lowBatteryThreshold < 0 ||
-        lowBatteryThreshold > 100)
-    ) {
-      return res.status(400).json({ error: 'lowBatteryThreshold must be a number between 0 and 100' });
-    }
-    // lowBatteryVoltageThreshold (mV) is optional (older clients omit it). MeshCore
-    // nodes report battery voltage; 0-20000 mV covers single-cell through multi-cell packs.
-    if (
-      lowBatteryVoltageThreshold !== undefined &&
-      (typeof lowBatteryVoltageThreshold !== 'number' ||
-        !Number.isFinite(lowBatteryVoltageThreshold) ||
-        lowBatteryVoltageThreshold < 0 ||
-        lowBatteryVoltageThreshold > 20000)
-    ) {
-      return res.status(400).json({ error: 'lowBatteryVoltageThreshold must be a number between 0 and 20000' });
-    }
-
-    // Validate appriseUrls is an array of strings if provided
-    if (appriseUrls !== undefined && !Array.isArray(appriseUrls)) {
-      return res.status(400).json({ error: 'appriseUrls must be an array' });
-    }
-    if (appriseUrls && appriseUrls.some((url: any) => typeof url !== 'string')) {
-      return res.status(400).json({ error: 'appriseUrls must be an array of strings' });
-    }
-
-    // Validate mutedChannels
-    if (mutedChannels !== undefined && !Array.isArray(mutedChannels)) {
-      return res.status(400).json({ error: 'mutedChannels must be an array' });
-    }
-    if (mutedChannels && mutedChannels.some((r: any) =>
-      typeof r !== 'object' || r === null ||
-      typeof r.channelId !== 'number' ||
-      (r.muteUntil !== null && typeof r.muteUntil !== 'number')
-    )) {
-      return res.status(400).json({ error: 'mutedChannels entries must have channelId (number) and muteUntil (number|null)' });
-    }
-
-    // Validate mutedDMs
-    if (mutedDMs !== undefined && !Array.isArray(mutedDMs)) {
-      return res.status(400).json({ error: 'mutedDMs must be an array' });
-    }
-    if (mutedDMs && mutedDMs.some((r: any) =>
-      typeof r !== 'object' || r === null ||
-      typeof r.nodeUuid !== 'string' ||
-      (r.muteUntil !== null && typeof r.muteUntil !== 'number')
-    )) {
-      return res.status(400).json({ error: 'mutedDMs entries must have nodeUuid (string) and muteUntil (number|null)' });
-    }
-
-    const prefs = {
-      enableWebPush,
-      enableApprise,
-      enabledChannels,
-      enableDirectMessages,
-      notifyOnEmoji,
-      notifyOnMqtt: notifyOnMqtt ?? true,
-      notifyOnNewNode,
-      notifyOnTraceroute,
-      notifyOnInactiveNode: notifyOnInactiveNode ?? false,
-      notifyOnLowBattery: notifyOnLowBattery ?? false,
-      lowBatteryThreshold: lowBatteryThreshold ?? 20,
-      lowBatteryVoltageThreshold: lowBatteryVoltageThreshold ?? 3300,
-      notifyOnServerEvents: notifyOnServerEvents ?? false,
-      prefixWithNodeName: prefixWithNodeName ?? false,
-      monitoredNodes: monitoredNodes ?? [],
-      whitelist,
-      blacklist,
-      appriseUrls: appriseUrls ?? [],
-      mutedChannels: mutedChannels ?? [],
-      mutedDMs: mutedDMs ?? [],
-    };
-
-    const success = await saveUserNotificationPreferencesAsync(userId, prefs, sourceId);
-
-    if (success) {
-      logger.info(
-        `✅ Saved notification preferences for user ${userId} source=${sourceId ?? '(default)'} (WebPush: ${enableWebPush}, Apprise: ${enableApprise})`
-      );
-      res.json({ success: true });
-    } else {
-      res.status(500).json({ error: 'Failed to save preferences' });
-    }
-  } catch (error: any) {
-    logger.error('Error saving notification preferences:', error);
-    res.status(500).json({ error: error.message || 'Failed to save preferences' });
-  }
-  }
-);
-
-// ==========================================
-// Apprise Notification Endpoints
-// ==========================================
-
-// Get Apprise status (admin only)
-apiRouter.get('/apprise/status', requireAdmin(), async (_req, res) => {
-  try {
-    const isAvailable = appriseNotificationService.isAvailable();
-    res.json({
-      available: isAvailable,
-      enabled: await databaseService.settings.getSetting('apprise_enabled') === 'true',
-      url: await databaseService.settings.getSetting('apprise_url') || 'http://localhost:8000',
-    });
-  } catch (error: any) {
-    logger.error('Error getting Apprise status:', error);
-    res.status(500).json({ error: error.message || 'Failed to get Apprise status' });
-  }
-});
-
-// Send test Apprise notification (admin only)
-apiRouter.post(
-  '/apprise/test',
-  requireAdmin(),
-  requirePermission('settings', 'write', { sourceIdFrom: 'body' }),
-  async (req, res) => {
-  try {
-    const userId = req.session?.userId;
-    if (!userId) {
-      return res.status(401).json({ success: false, message: 'Not authenticated' });
-    }
-
-    const sourceId = typeof req.body?.sourceId === 'string' && req.body.sourceId
-      ? req.body.sourceId
-      : undefined;
-    if (!sourceId) {
-      return res.status(400).json({ success: false, message: 'sourceId is required' });
-    }
-
-    // Resolve source for sourceName
-    const source = await databaseService.sources.getSource(sourceId);
-    if (!source) {
-      return res.status(400).json({ success: false, message: `Unknown sourceId: ${sourceId}` });
-    }
-
-    // Get user's Apprise URLs from their preferences (per-source)
-    const prefs = await getUserNotificationPreferencesAsync(userId, sourceId);
-    if (!prefs || !prefs.appriseUrls || prefs.appriseUrls.length === 0) {
-      return res.json({
-        success: false,
-        message: 'No Apprise URLs configured in your notification preferences',
-      });
-    }
-
-    // Get local node name for prefix
-    const localNodeInfo = meshtasticManager.getLocalNodeInfo();
-    const localNodeName = localNodeInfo?.longName || null;
-
-    // Apply prefix if user has it enabled
-    const baseBody = 'This is a test notification from MeshMonitor via Apprise';
-    const body = await applyNodeNamePrefixAsync(userId, baseBody, localNodeName);
-
-    // Send to user's configured URLs
-    const success = await appriseNotificationService.sendNotificationToUrls(
-      {
-        title: 'Test Notification',
-        body,
-        type: 'info',
-        sourceId,
-        sourceName: source.name ?? sourceId,
-      },
-      prefs.appriseUrls
-    );
-
-    if (success) {
-      res.json({ success: true, message: 'Test notification sent successfully' });
-    } else {
-      res.json({ success: false, message: 'Failed to send notification - check your Apprise URLs' });
-    }
-  } catch (error: any) {
-    logger.error('Error sending test Apprise notification:', error);
-    res.status(500).json({ error: error.message || 'Failed to send test notification' });
-  }
-  }
-);
-
-// Get configured Apprise URLs (admin only)
-apiRouter.get('/apprise/urls', requireAdmin(), async (_req, res) => {
-  try {
-    const configFile = process.env.APPRISE_CONFIG_DIR
-      ? `${process.env.APPRISE_CONFIG_DIR}/urls.txt`
-      : '/data/apprise-config/urls.txt';
-
-    // Check if file exists
-    const fs = await import('fs/promises');
-    try {
-      const content = await fs.readFile(configFile, 'utf-8');
-      const urls = content
-        .split('\n')
-        .map(line => line.trim())
-        .filter(line => line.length > 0 && !line.startsWith('#'));
-
-      res.json({ urls });
-    } catch (error: any) {
-      // File doesn't exist or can't be read - return empty array
-      if (error.code === 'ENOENT') {
-        res.json({ urls: [] });
-      } else {
-        throw error;
-      }
-    }
-  } catch (error: any) {
-    logger.error('Error reading Apprise URLs:', error);
-    res.status(500).json({ error: error.message || 'Failed to read Apprise URLs' });
-  }
-});
-
-// Configure Apprise URLs (admin only)
-apiRouter.post('/apprise/configure', requireAdmin(), async (req, res) => {
-  try {
-    const { urls } = req.body;
-
-    if (!Array.isArray(urls)) {
-      return res.status(400).json({ error: 'URLs must be an array' });
-    }
-
-    // Security: Validate URL schemes to prevent malicious URLs
-    // Comprehensive list of all Apprise-supported notification services
-    // Reference: https://github.com/caronc/apprise
-    const ALLOWED_SCHEMES = [
-      // Core Apprise
-      'apprise',
-      'apprises',
-
-      // Chat & Messaging
-      'discord',
-      'slack',
-      'msteams',
-      'teams',
-      'guilded',
-      'revolt',
-      'matrix',
-      'matrixs',
-      'mmost',
-      'mmosts',
-      'rocket',
-      'rockets',
-      'ryver',
-      'zulip',
-      'twist',
-      'gchat',
-      'flock',
-
-      // Instant Messaging & Social
-      'telegram',
-      'tgram',
-      'signal',
-      'signals',
-      'whatsapp',
-      'line',
-      'mastodon',
-      'mastodons',
-      'misskey',
-      'misskeys',
-      'bluesky',
-      'reddit',
-      'twitter',
-
-      // Team Communication
-      'workflows',
-      'wxteams',
-      'wecombot',
-      'feishu',
-      'lark',
-      'dingtalk',
-
-      // Push Notifications
-      'pushover',
-      'pover',
-      'pushbullet',
-      'pbul',
-      'pushed',
-      'pushme',
-      'pushplus',
-      'pushdeer',
-      'pushdeers',
-      'pushy',
-      'prowl',
-      'simplepush',
-      'spush',
-      'popcorn',
-      'push',
-
-      // Notification Services
-      'ntfy',
-      'ntfys',
-      'gotify',
-      'gotifys',
-      'join',
-      'ifttt',
-      'notica',
-      'notifiarr',
-      'notifico',
-      'onesignal',
-      'kumulos',
-      'bark',
-      'barks',
-      'chanify',
-      'serverchan',
-      'schan',
-      'qq',
-      'wxpusher',
-
-      // Incident Management & Monitoring
-      'pagerduty',
-      'pagertree',
-      'opsgenie',
-      'spike',
-      'splunk',
-      'victorops',
-      'signl4',
-
-      // Email Services
-      'mailto',
-      'email',
-      'smtp',
-      'smtps',
-      'ses',
-      'mailgun',
-      'sendgrid',
-      'smtp2go',
-      'sparkpost',
-      'o365',
-      'resend',
-      'sendpulse',
-
-      // SMS Services
-      'bulksms',
-      'bulkvs',
-      'burstsms',
-      'clickatell',
-      'clicksend',
-      'd7sms',
-      'freemobile',
-      'httpsms',
-      'atalk',
-
-      // Cloud/IoT/Home
-      'fcm',
-      'hassio',
-      'hassios',
-      'homeassistant',
-      'parsep',
-      'parseps',
-      'aws',
-      'sns',
-
-      // Media Centers
-      'kodi',
-      'kodis',
-      'xbmc',
-      'xbmcs',
-      'emby',
-      'embys',
-      'enigma2',
-      'enigma2s',
-
-      // Collaboration & Productivity
-      'ncloud',
-      'nclouds',
-      'nctalk',
-      'nctalks',
-      'office365',
-
-      // Streaming & Gaming
-      'streamlabs',
-      'strmlabs',
-
-      // Specialized
-      'lametric',
-      'synology',
-      'synologys',
-      'vapid',
-      'mqtt',
-      'mqtts',
-      'rsyslog',
-      'syslog',
-      'dapnet',
-      'aprs',
-      'growl',
-      'pjet',
-      'pjets',
-      'psafer',
-      'psafers',
-      'spugpush',
-      'pushsafer',
-
-      // Generic webhooks & protocols
-      'webhook',
-      'webhooks',
-      'json',
-      'xml',
-      'form',
-      'http',
-      'https',
-    ];
-
-    const invalidUrls: string[] = [];
-    const validUrls = urls.filter((url: string) => {
-      if (typeof url !== 'string' || !url.trim()) {
-        invalidUrls.push(url);
-        return false;
-      }
-
-      // Extract scheme using regex instead of URL parser
-      // This allows Apprise URLs with special characters (colons, multiple slashes, etc.)
-      // that don't conform to strict URL syntax but are valid for Apprise
-      // Support both "scheme://" format and special cases like "mailto:"
-      const schemeMatch = url.match(/^([a-zA-Z][a-zA-Z0-9+.-]*):/);
-
-      if (!schemeMatch) {
-        invalidUrls.push(url);
-        return false;
-      }
-
-      const scheme = schemeMatch[1].toLowerCase();
-
-      if (!ALLOWED_SCHEMES.includes(scheme)) {
-        invalidUrls.push(url);
-        return false;
-      }
-
-      return true;
-    });
-
-    if (invalidUrls.length > 0) {
-      return res.status(400).json({
-        error: 'Invalid or disallowed URL schemes detected',
-        invalidUrls,
-        allowedSchemes: ALLOWED_SCHEMES,
-      });
-    }
-
-    const result = await appriseNotificationService.configureUrls(validUrls);
-    res.json(result);
-  } catch (error: any) {
-    logger.error('Error configuring Apprise URLs:', error);
-    res.status(500).json({ error: error.message || 'Failed to configure Apprise URLs' });
-  }
-});
-
-// Enable/disable Apprise system-wide (admin only)
-apiRouter.put('/apprise/enabled', requireAdmin(), async (req, res) => {
-  try {
-    const { enabled } = req.body;
-
-    if (typeof enabled !== 'boolean') {
-      return res.status(400).json({ error: 'Enabled must be a boolean' });
-    }
-
-    await databaseService.settings.setSetting('apprise_enabled', enabled ? 'true' : 'false');
-    logger.info(`✅ Apprise ${enabled ? 'enabled' : 'disabled'} system-wide`);
-    res.json({ success: true, enabled });
-  } catch (error: any) {
-    logger.error('Error updating Apprise enabled status:', error);
-    res.status(500).json({ error: error.message || 'Failed to update Apprise status' });
-  }
-});
 
 // Serve static files from the React app build
 const buildPath = path.join(__dirname, '../../dist');
